Adds UpdateBoardController
