--- src/main/java/tttweb/controllers/NewMoveController.java
@@ -1,4 +1,0 @@
-package tttweb.controllers;
-
-public class NewMoveController {
-}
